Refactor: Separate group_repository and membership_repository
